staging: refactor completion for clearer scoping
